Creación y corrección de componentes navbar y footer
