+++ .firebase/hosting.YnVpbGQ.cache
@@ -1,16 +1,16 @@
 robots.txt,1755810394978,b2090cf9761ef60aa06e4fab97679bd43dfa5e5df073701ead5879d7c68f1ec5
 manifest.json,1755810394977,4368aeaf848ae2e048765562c289452f33ad2a175c4b1951ea8bdf2ada0d5b10
+logo512.png,1755810394977,212b102aa09e51b3b3e06647e81f7801a61333e171f6582e8124379aabccb41d
 logo192.png,1755810394977,79e2b749561016bc8af300ea19f48347ceed3cb1a54f48ae456172eca45e08f0
 favicon.ico,1755810394976,27edce7be5922cf0bef7d4136f69b5bfbdd5bf8c13c7b026f71187d41a00aa7d
-index.html,1756397139665,5b36e4014cb0abb5bc921640b66232ec55e9724eedb4c8183ae4a79d137354a4
-asset-manifest.json,1756397139666,ed444d747065d66fa764f7a3426e7d85798057f7ca1fa548cf23a507e368728b
-static/js/main.34ad8317.js.LICENSE.txt,1756397139676,110427dedf71d596f811b001fda37d713ec83466481c97186fe871e6d2dc46c4
-static/js/453.3acde13a.chunk.js,1756397139676,15e62cddd0eee15ce4669664e4a4c541fe8c735c7f5951be6b38454011ec2d6d
-static/js/453.3acde13a.chunk.js.map,1756397139676,2b66cd61c97e08d70ba17afacc15a854df2adc6a33eb213d44b29bf8399ca1e0
-static/media/user.174cd90b5a7f1d66f9bc.png,1756397139676,511ab45d442d2a92598177afb9a84d4806371c05bd967d12893a346aaa736fe3
-logo512.png,1755810394977,212b102aa09e51b3b3e06647e81f7801a61333e171f6582e8124379aabccb41d
-static/css/main.69a3a12b.css,1756397139676,e9c507e4c1a6c59d9220c6160753fc67bacf63955112afe776e91357c9e19b4c
-static/media/brilla.478619515c4a57ad13ff.png,1756397139675,f4cf77740dc9d82792e49fcbb3ea05d439c641338bd48a15034db4887fb8f189
-static/css/main.69a3a12b.css.map,1756397139676,b1174f6792a3e120bcd22449486a14094e995d80792918259da6bccf904f4897
-static/js/main.34ad8317.js,1756397139676,bcf95f7a495b0fd183677cc8319b26a3f2dcc094d50762ccce193555e8c67d50
-static/js/main.34ad8317.js.map,1756397139692,7317d6fdd9751060e948f1e8b1dd7d819cdfbbc8b76a70ea847a702ea159b751
+index.html,1756427358672,095b5b2febffd2e5f75555b5a011815bf1758394c88af0ad792f9b1a7a641c1a
+asset-manifest.json,1756427358673,ed444d747065d66fa764f7a3426e7d85798057f7ca1fa548cf23a507e368728b
+static/js/main.34ad8317.js.LICENSE.txt,1756427358689,110427dedf71d596f811b001fda37d713ec83466481c97186fe871e6d2dc46c4
+static/js/453.3acde13a.chunk.js,1756427358689,15e62cddd0eee15ce4669664e4a4c541fe8c735c7f5951be6b38454011ec2d6d
+static/js/453.3acde13a.chunk.js.map,1756427358689,2b66cd61c97e08d70ba17afacc15a854df2adc6a33eb213d44b29bf8399ca1e0
+static/media/user.174cd90b5a7f1d66f9bc.png,1756427358689,511ab45d442d2a92598177afb9a84d4806371c05bd967d12893a346aaa736fe3
+static/css/main.69a3a12b.css,1756427358689,e9c507e4c1a6c59d9220c6160753fc67bacf63955112afe776e91357c9e19b4c
+static/media/brilla.478619515c4a57ad13ff.png,1756427358689,f4cf77740dc9d82792e49fcbb3ea05d439c641338bd48a15034db4887fb8f189
+static/css/main.69a3a12b.css.map,1756427358689,b1174f6792a3e120bcd22449486a14094e995d80792918259da6bccf904f4897
+static/js/main.34ad8317.js,1756427358689,bcf95f7a495b0fd183677cc8319b26a3f2dcc094d50762ccce193555e8c67d50
+static/js/main.34ad8317.js.map,1756427358691,7317d6fdd9751060e948f1e8b1dd7d819cdfbbc8b76a70ea847a702ea159b751
