+++ src/pages/components/Footer.js
@@ -1,0 +1,11 @@
+const Footer = () => {
+  return (
+    <footer className="footer mt-auto">
+      <div className="container">
+        <small>© 2025 Brilla. All rights reserved.</small>
+      </div>
+    </footer>
+  )
+}
+
+export default Footer
