+++ src/pages/components/Spinner.js
